+++ src/Application/Query/Happening/Webinar/GetWebinarViewQueryHandler.php
@@ -39,6 +39,7 @@ class GetWebinarViewQueryHandler
         );
 
         $sessionId = $session->getSessionId();
+        $timeRemainingInSeconds = max(0, $happening->getEnd()->getTimestamp() - $this->dateTime->getTimestamp());
 
         return new WebinarView(
             $happening->getTitle($query->getLocale()),
@@ -48,6 +49,8 @@ class GetWebinarViewQueryHandler
             $isSpeaker,
             new TimeRangeView($happening->getBegin(), $happening->getEnd()),
             $this->dateTime,
+            $timeRemainingInSeconds,
+            round($timeRemainingInSeconds * 0.2),
             $happening->getWebinarHeaderImage($query->getLocale())
         );
     }