+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/VideoConference/Webinar.js
@@ -16,9 +16,10 @@ function Webinar(element, isSpeaker) {
     this.token = element.getAttribute('data-token');
     this.sessionId = element.getAttribute('data-session-id');
     this.apiKey = element.getAttribute('data-api-key');
-    this.webinarEndTime = element.getAttribute('data-webinar-end-time');
-    this.webinarStartTime = element.getAttribute('data-webinar-start-time');
-    this.currentTime = element.getAttribute('data-current-time');
+
+    this.timeRemaining = element.getAttribute('data-time-remaining');
+    this.warningRemainingTime = element.getAttribute('data-warning-time-remaining');
+
     this.chatWaitingMessage = element.getAttribute('data-chat-waiting-message');
     this.userCompleteName = element.getAttribute('data-user-complete-name');
     this.helperContainer = element.querySelector('.video-helper');
@@ -58,10 +59,11 @@ function Webinar(element, isSpeaker) {
     }
     this.timerContainer = element.querySelector('.timer');
     this.countDownContainer = element.querySelector('.timer span.countdown');
+    this.viewersCount = 0;
+    this.viewersContainer = element.querySelector('.viewers-container');
+    this.viewersTextContainer = element.querySelector('.viewers');
 
     if (this.isSpeaker) {
-        this.viewersCount = 0;
-        this.viewersContainer = element.querySelector('.viewers');
         this.thereIsAlreadyAScreenShareInProgressMessage = element.getAttribute('data-screen-share-already-in-progress-message');
 
         this.endScreenSharingButton = element.querySelector('#end-screensharing');
@@ -131,7 +133,7 @@ Webinar.prototype.init = function () {
         return;
     }
 
-    this.session = TokboxInstance.initSession(this.apiKey, this.sessionId, {connectionEventsSuppressed: !this.isSpeaker});
+    this.session = TokboxInstance.initSession(this.apiKey, this.sessionId);
 
     this.session.on('streamCreated', function (event) {
         this.hideElement(this.helperContainer);
@@ -194,10 +196,7 @@ Webinar.prototype.init = function () {
 };
 
 Webinar.prototype.updateViewers = function () {
-    // remove speaker
-    const viewersCount = this.viewersCount - 1;
-
-    this.viewersContainer.textContent = '' + viewersCount;
+    this.viewersTextContainer.textContent = this.viewersCount;
 };
 
 /**
@@ -209,6 +208,8 @@ Webinar.prototype.connect = function () {
         this.showElement(this.toggleAudioElement);
         this.showElement(this.toggleVideoElement);
         this.showElement(this.startScreenSharingButton);
+        this.showElement(this.timerContainer);
+        this.showElement(this.viewersContainer);
 
         if (!error) {
             if (this.isSpeaker) {
@@ -275,15 +276,21 @@ Webinar.prototype.publishStream = function () {
  * Disconnect from the session
  */
 Webinar.prototype.disconnect = function () {
-    this.session.disconnect();
-    this.session.off();
+    if (this.session) {
+        this.session.disconnect();
+        this.session.off();
+    }
+
     this.session = null;
 
     if (window.opener) {
         window.opener.location.reload(true);
+        window.close();
+
+        return;
     }
 
-    window.close();
+    window.history.go(-1);
 };
 
 Webinar.prototype.handlePublish = function (error) {
@@ -510,7 +517,7 @@ Webinar.prototype.countDownBeforeEnd = function () {
         return;
     }
 
-    new Counter(this.webinarStartTime, this.webinarEndTime, this.currentTime, this.countDownContainer, this.timerContainer);
+    new Counter(parseInt(this.timeRemaining, 10), parseInt(this.warningRemainingTime, 10), this.countDownContainer, this.timerContainer);
 };
 
 Webinar.prototype.hidePublisher = function () {